Add third line of text to hello_world.txt
Added another line of text to test how to change the Git commit message
editor to VSCode. Before this commit, I was unsure how to create a body
commit message, moreover how to write a commit in VSCode. This was
disadvantageous because writing Git commit messages with bodies are not
easy to write in the shell with the -m option.

Now if I want to write a Git commit message with a body, I just execute
`git commit` in shell, which tells Git to open VSCode as the text editor
and provide more information about the commit. After typing the commit
message, I can save it and exit the tab in VSCode.

To address this issue, I executed the git command:

`git config --global core.editor "code --wait"`

The `git config` command tells Git to specify Git configuration
settings. The `--global core.editor` option configures the default text
editor. The `"code --wait"` argument tells Git to use VSCode as the
default text editor and not to proceed with the command until you are
done with the editing task in VSCode.